move getFileName and utilise it in compare model
