+++ packages/client/hmi-client/src/components/workflow/ops/compare-datasets/tera-compare-datasets-drilldown.vue
@@ -111,6 +111,7 @@
 											v-model="data[dataset.id]"
 											:options="
 												dataset.columns
+													?.filter((ele) => ele.fileName === getFileName(dataset))
 													?.map((ele) => ele.name)
 													.filter(
 														(ele) =>
@@ -401,6 +402,7 @@ import { useCharts, type ChartData } from '@/composables/useCharts';
 import { DataArray } from '@/services/models/simulation-service';
 import { mean, stddev, quantile } from '@/utils/stats';
 import { displayNumber } from '@/utils/number';
+import { getFileName } from '@/services/dataset';
 import TeraCriteriaOfInterestCard from './tera-criteria-of-interest-card.vue';
 import {
 	blankCriteriaOfInterest,
@@ -425,6 +427,7 @@ const compareOptions: { label: string; value: CompareValue }[] = [
 ];
 
 const datasets = ref<Dataset[]>([]);
+console.log(datasets);
 const datasetResults = ref<{
 	results: DataArray[];
 	summaryResults: DataArray[];
