+++ packages/client/hmi-client/src/components/workflow/scenario-templates/calibrate-ensemble/tera-calibrate-ensemble-template.vue
@@ -154,8 +154,7 @@ import Button from 'primevue/button';
 import TabView from 'primevue/tabview';
 import TabPanel from 'primevue/tabpanel';
 import TeraInputText from '@/components/widgets/tera-input-text.vue';
-import { getFileName } from '@/services/calibrate-workflow';
-import { getDataset } from '@/services/dataset';
+import { getDataset, getFileName } from '@/services/dataset';
 import CalibrateEnsembleImg from '@/assets/svg/template-images/calibrate-ensemble-model-thumbnail.svg';
 import { ScenarioHeader } from '../base-scenario';
 import { CalibrateEnsembleScenario } from './calibrate-ensemble-scenario';
