+++ packages/client/hmi-client/src/services/dataset.ts
@@ -282,6 +282,10 @@ const mergeResults = (...results: DataArray[]) => {
 	return result;
 };
 
+const getFileName = (dataset: Dataset) =>
+	// If our dataset includes a result.csv we will ensure to pick it.
+	dataset.fileNames?.includes('result.csv') ? 'result.csv' : (dataset?.fileNames?.[0] ?? '');
+
 export {
 	getDataset,
 	updateDataset,
@@ -296,5 +300,6 @@ export {
 	getRawContent,
 	getCsvAsset,
 	getDatasetResultCSV,
-	mergeResults
+	mergeResults,
+	getFileName
 };
